separated definitions from intended script
